moved Equipment.php into modes folder
